feat(sites): allow decimal area_sqft (widen integer column to numeric)
sites.area_sqft was integer, so fractional square footage (e.g. 1120.50) typed
at New Pipeline / Add Details was truncated. Widen it to numeric(12,2) to match
the sibling *_area_sqft columns.

- Migration 20260801_widen_area_sqft_to_numeric.sql: ALTER COLUMN ... TYPE
  numeric(12,2) USING area_sqft::numeric, guarded so a re-run is a no-op. Runs
  once via the schema_migrations ledger runner. Verified read-only against the
  live DB: 218 rows, 0 nulls, all values cast cleanly (max 34343.00).
- Model area_sqft -> Numeric(12,2); SaveDetailsRequest/SiteResponse int -> float;
  bd_service coerces the field with _to_float (removed the now-unused _to_int).
- Frontend already sends decimals (New Pipeline input step="any"; Number()).

+++ backend/app/domain/schemas/site.py
@@ -155,7 +155,7 @@ class SaveDetailsRequest(BaseModel):
     tenure: Optional[int] = None
     total_op_cost: Optional[float] = Field(default=None, validation_alias=AliasChoices("total_op_cost", "totalOpCost"))
     # Pipeline-stage area (sqft) — editable via Add Details
-    area_sqft: Optional[int] = Field(default=None, ge=0, validation_alias=AliasChoices("area_sqft", "areaSqft"))
+    area_sqft: Optional[float] = Field(default=None, ge=0, validation_alias=AliasChoices("area_sqft", "areaSqft"))
     # Staggered escalation schedule — passed through to sites row
     staggered_escalation: Optional[List[StaggeredEscalationItem]] = Field(
         default=None,
@@ -220,7 +220,7 @@ class SiteResponse(BaseModel):
     expected_escalation_pct: Optional[float] = None
     expected_escalation_years: Optional[int] = None
     expected_revshare_pct: Optional[float] = None
-    area_sqft: int = 0
+    area_sqft: float = 0
     staggered_escalation: Optional[list] = None
 
     @field_validator("staggered_escalation", mode="before")
